fix zone so that it no longer breaks randomly

+++ src/z_zone.h
@@ -45,22 +45,22 @@
 
 
 void	Z_Init (void);
-void*	Z_Malloc (long size, long tag, void *ptr);
+void*	Z_Malloc (int size, int tag, void *ptr);
 void    Z_Free (void *ptr);
-void    Z_FreeTags (long lowtag, long hightag);
-void    Z_DumpHeap (long lowtag, long hightag);
+void    Z_FreeTags (int lowtag, int hightag);
+void    Z_DumpHeap (int lowtag, int hightag);
 void    Z_FileDumpHeap (FILE *f);
 void    Z_CheckHeap (void);
-void    Z_ChangeTag2 (void *ptr, long tag);
-long     Z_FreeMemory (void);
+void    Z_ChangeTag2 (void *ptr, int tag);
+int     Z_FreeMemory (void);
 
 
 typedef struct memblock_s
 {
-    long			size;	// including the header and possibly tiny fragments
+    int			size;	// including the header and possibly tiny fragments
     void**		user;	// NULL if a free block
-    long			tag;	// purgelevel
-    long			id;	// should be ZONEID
+    int			tag;	// purgelevel
+    int			id;	// should be ZONEID
     struct memblock_s*	next;
     struct memblock_s*	prev;
 } memblock_t;
@@ -72,7 +72,7 @@ typedef struct memblock_s
 #define Z_ChangeTag(p,t) \
 { \
       if (( (memblock_t *)( (byte *)(p) - sizeof(memblock_t)))->id!=0x1d4a11) \
-	  I_Error("Z_CT at "__FILE__":%li",__LINE__); \
+	  I_Error("Z_CT at "__FILE__":%i",__LINE__); \
 	  Z_ChangeTag2(p,t); \
 };
